added PythonTest interface
